Add acceleration values to TSV output.
We grabbed the acceleration values for all three axes correctly, but
did not include them in the properties in the TSV output. This adds
them in the correct place, right after the engine RPM values.

Also added a utility function for writing a tab-separated list of
column names for use as a file header, which will be added later.

Closes issue #1