Add CSV support and rename file output script.
We add support for Excel-style CSV files by using Python's csv library.
Switching the output is controlled with a new command line parameter.
Because the script no longer supports just TSVs, the script was renamed
to data2file.py.

The default behaviour is now to overwrite the file by default, so there
is a command line parameter added for appending to an existing file.

The ForzaDataPacket class is changed to contain two class variables
and a class method, as the format specifier and list of property names
are both constant across all objects.