Add support for the dash cam format and update readme
With the September 2018 update to Forza Motorsport 7 came also
an update to the data output feature. A new format called
"dash cam" arrived, with a number of new features. This patch
adds support for that format to the ForzaDataPacket class,
as well as to the utility script. The utility script has
also changed to expect the new format as default. Because
the new format also contains all the old format's features,
it seems reasonable to expect users to want to only use
the new format. However, the old format is still supported.

The readme file is updated to document how to specify the
old format for the utility script, and also to let everyone
know that contributions are welcome.