Add support for configuration file and system time.
The data logging utility can now read a YAML configuration file.
Said file can override all the command line options. It can also
contain a list of specific data to log. The format of the file and
the potential options is described in a new markdown document.

Adds wall_clock as a potential parameter to log, which is the
local time. This is only available through the configuration file.

Renamed three packet properties, they were poorly named in the
specification. Best lap time, last lap time, and current lap time
now all reflect that they are time measurements (in seconds).

Added some additional output when in verbose mode so it's easy
to see that data is being logged.

This closes issues #3 and #4.